+++ dune/functions/functionspacebases/constrainedglobalbasis.hh
@@ -6,6 +6,8 @@
 #include <dune/common/reservedvector.hh>
 #include <dune/common/typeutilities.hh>
 #include <dune/common/concept.hh>
+#include <dune/common/deprecated.hh>
+#include <dune/common/version.hh>
 
 #include <dune/functions/common/type_traits.hh>
 #include <dune/functions/functionspacebases/constrainedlocalindexset.hh>
@@ -138,7 +140,11 @@ public:
   }
 
   //! Return local index set for basis
-  LocalIndexSet localIndexSet() const
+  LocalIndexSet
+#if DUNE_VERSION_NEWER(DUNE_FUNCTIONS,2,7)
+    DUNE_DEPRECATED_MSG("localIndexSet() is deprecated. The indexing method are now available in the ConstrainedLocalView. Indices are computed when binding the ConstrainedLocalView.")
+#endif
+    localIndexSet() const
   {
     return LocalIndexSet(preBasis_.template indexSet<PrefixPath>());
   }